s3-integration init delete paths
