+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-core/src/main/java/de/muenchen/oss/digiwf/s3/integration/adapter/in/streaming/MessageProcessor.java
@@ -3,18 +3,18 @@ package de.muenchen.oss.digiwf.s3.integration.adapter.in.streaming;
 import de.muenchen.oss.digiwf.message.process.api.error.BpmnError;
 import de.muenchen.oss.digiwf.message.process.api.error.IncidentError;
 import de.muenchen.oss.digiwf.s3.integration.adapter.in.rest.mapper.PresignedUrlMapper;
-import de.muenchen.oss.digiwf.s3.integration.application.port.in.CreatePresignedUrlsInPort;
-import de.muenchen.oss.digiwf.s3.integration.application.port.in.FileExistenceException;
-import de.muenchen.oss.digiwf.s3.integration.application.port.in.FileSystemAccessException;
+import de.muenchen.oss.digiwf.s3.integration.application.port.in.*;
 import de.muenchen.oss.digiwf.s3.integration.application.port.out.IntegrationOutPort;
 import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.springframework.messaging.Message;
 
 import java.util.Map;
 import java.util.function.Consumer;
 
+@Slf4j
 @RequiredArgsConstructor
 public class MessageProcessor {
     /**
@@ -25,6 +25,8 @@ public class MessageProcessor {
     private static final String FILE_DOES_NOT_EXIST_ERROR_CODE = "FILE_DOES_NOT_EXIST_ERROR";
 
     private final CreatePresignedUrlsInPort createPresignedUrlsInPort;
+    private final FolderOperationsInPort folderOperationsInPort;
+    private final FileOperationsInPort fileOperationsInPort;
     private final IntegrationOutPort integration;
     private final PresignedUrlMapper presignedUrlMapper;
 
@@ -48,6 +50,30 @@ public class MessageProcessor {
                 integration.handleBpmnError(headers, new BpmnError(FILE_DOES_NOT_EXIST_ERROR_CODE, fee.getMessage()));
             } catch (FileSystemAccessException sae) {
                 integration.handleIncident(headers, new IncidentError(sae.getMessage()));
+            }
+        };
+    }
+
+    public Consumer<Message<PathsDTO>> deletePaths() {
+        return message -> {
+            try {
+                val payload = message.getPayload();
+                log.info("Delete paths request: {}", payload);
+                for (String path : payload.getPathsAsList()) {
+                    final String fullPath = payload.getFileContext() + "/" + path;
+                    if (fullPath.endsWith("/")) {
+                        this.folderOperationsInPort.deleteFolder(fullPath);
+                    } else {
+                        this.fileOperationsInPort.deleteFile(fullPath);
+                    }
+                }
+                integration.correlateProcessMessage(message.getHeaders(), Map.of());
+            } catch (ConstraintViolationException cve) {
+                integration.handleBpmnError(message.getHeaders(), new BpmnError(VALIDATION_ERROR_CODE, cve.getMessage()));
+            } catch (FileExistenceException fee) {
+                integration.handleBpmnError(message.getHeaders(), new BpmnError(FILE_DOES_NOT_EXIST_ERROR_CODE, fee.getMessage()));
+            } catch (FileSystemAccessException sae) {
+                integration.handleIncident(message.getHeaders(), new IncidentError(sae.getMessage()));
             }
         };
     }
